# CUDA performance optimization: asynchronous computation by using only one cudaStream