On master a performance bottleneck for GPU token generation is currently that the GPU has to wait between computations for the CPU to provide the next task. This PR queues many tasks at once, resulting in higher performance. For the following number PP is prompt processing and TG 128/1024 is the generation of 128/1024 tokens from an empty prompt:

| GPU                    | Test    | Model    | t/s master | t/s PR | Speedup |
|------------------------|---------|----------|------------|--------|---------|
| RTX 3090               | PP      | 7b q4_0  |       1000 |   1042 |    1.04 |
| RTX 3090               | PP      | 13b q4_0 |        658 |    667 |    1.01 |
| RTX 3090               | PP      | 33b q4_0 |        297 |    299 |    1.01 |
| RTX 3090               | TG 128  | 7b q4_0  |      63.69 |  84.67 |    1.33 |
| RTX 3090               | TG 128  | 13b q4_0 |      38.52 |  46.75 |    1.21 |
| RTX 3090               | TG 128  | 33b q4_0 |      17.70 |  19.91 |    1.13 |
| RTX 3090               | TG 1024 | 7b q4_0  |      58.48 |  72.99 |    1.25 |
| RTX 3090               | TG 1024 | 13b q4_0 |      35.98 |  41.81 |    1.16 |
| RTX 3090               | TG 1024 | 33b q4_0 |      16.50 |  18.46 |    1.12 |
| GTX 1070               | PP      | 7b q4_0  |        310 |    309 |    1.00 |
| GTX 1070               | TG 128  | 7b q4_0  |      15.92 |  16.68 |    1.05 |
| GTX 1070               | TG 1024 | 7b q4_0  |      15.14 |  15.90 |    1.05 |
| GTX 1070 + GTX 1050 ti | PP      | 7b q4_0  |        103 |    100 |    0.97 |
| GTX 1070 + GTX 1050 ti | TG 128  | 7b q4_0  |      19.03 |  19.57 |    1.03 |
| GTX 1070 + GTX 1050 ti | TG 1024 | 7b q4_0  |      17.96 |  18.84 |    1.03 |

The implementation works by reducing the number of CUDA streams per GPU from 16 to 1. That way the CPU can evaluate the entire ggml compute graph and queue all kernels with a guarantee that the GPU will wait for the previous kernel to finish before starting with the next one; `cudaDeviceSynchronize` is only called rarely when e.g. data needs to be copied from the VRAM to RAM. Theoretically you could instead e.g. use a bunch of CUDA events to do something similar but I found that this adds significant overhead.

As a side effect reducing the number of CUDA streams to 1 causes cuBLAS to become deterministic.

Multi GPU performance is only slightly improved because the splitting of data between GPUs requires synchronization; I may be able to at some point come up with a better way to do this.