@shouyiwang 2023-06-17
I merged this PR into the master branch, which includes ikawrakow's Cuda K-quant optimization. 
It achieved a remarkable performance of 28 ms/token on the 33b q4_K_M using a 4090.  And the GPU utilization remains consistently high at around 92%. 
The performance is truly fantastic!