@shouyiwang 2023-06-17
I have noticed that the VRAM usage has decreased by around 10% since the initial implementation of KV cache in the GPU. Looks like another miracle. What happened?
It has made the "--low-vram" option less relevant as it can only slightly reduce VRAM usage since there isn't much to offload to CPU.