@JohannesGaessler 2023-06-17
I don't see why the VRAM usage would have decreased.