@shouyiwang 2023-06-17
@JohannesGaessler 
I have just reviewed all the CUDA performance-related commits and conducted a comparison. 
For a 33b Q4_K_M model running on a single 4090 GPU, the VRAM usage and speed are as follows:

| Status                     | Memory    | Time/Token (ms/token) | Concurrency |
| -------------------------- | --------- | --------------------- | ----------- |
| Before #1827               | 21366MB   | 55                    | -t 6        |
| After #1827                | 22142MB   | 38                    | -t 1        |
| After #1862                | 22142MB   | 33.5                  | -t 1       |
| After: #1891               | 22142MB   | 33.2                  | -t 1       |
| Master with this PR        | 22142MB   | 28                    | -t 1        |

It appears that there had been several changes related to VRAM usage in #1827. Previously, I was unable to fit the entire 33b Q4_K_M model into 24GB of VRAM with the earlier version of the PR. Now I have 2GB of headroom remaining.